refactor(frontend): 残存3 Screen を feature hook に分離（#104-1）
ProductMaster / AdminStats / TotpSettings の API・service 呼び出しを
features 配下の hook に移し、Screen を規約どおり 150 行以内に整理する。
