+++ frontend/src/features/admin/index.ts
@@ -1,3 +1,4 @@
 export { usePromptEditor } from './hooks/usePromptEditor';
+export { useAdminStats } from './hooks/useAdminStats';
 export { PromptEditorForm } from './components/PromptEditorForm';
 export { PromptEditorList } from './components/PromptEditorList';
