+++ frontend/src/features/productMaster/index.ts
@@ -1,0 +1,1 @@
+export { useProductMaster } from './hooks/useProductMaster';
